Add Error and SetError to all datasets.
This addresses https://github.com/doug-martin/goqu/issues/150 by
adding SetError and Error methods to each dataset. This will allow
end-user query building (e.g. users of goqu) to be able to mark a
dataset as being in a failed state via SetError and not have to track
that separately.

It only sets the error if one has not already been set. It also sets it
on the underlying SQLBuilder so that it too will know about the error
and not try to proceed with any futher query building. As such, future
calls to ToSQL will return the error as well.