Do not interpolate literals when prepared=true.
This addresses https://github.com/doug-martin/goqu/issues/132

It is a reset of closed PR: https://github.com/doug-martin/goqu/pull/151

The gist of this fix is to modify the builders to not interpolate
nil and boolean values so that prepared statements can work properly.
This makes prepared=true work more consistently across all types.

The scope of this change:

(1) Modify literalNil to use a placeholder if prepared=true similar to
literalBool and literalInt

(2) Modify checkBoolExpType to not turn (in)equality on nil or bool
values to an IS expression.

(3) LOTS of tests and examples updated to reflect the new behavior.

(4) Fix bugs from https://github.com/doug-martin/goqu/pull/151 found by
@doug-martin to preserve correct behavior with Is/IsNot and not
incorrectly map them to Eq/NotEq.