feat: add SetDefaultPrepared which controls query value interpolation
This switches the datasets' internal isPrepared bools into a custom type
that resolves back into a bool when passed into the SQL builder. This
could have also been a *bool but I wanted to avoid nil checks and
potential sources of confusion if e.g. a child dataset happened to
dereference and mutate the pointer's value.